+++ soam/workflow/time_series_extractor.py
@@ -268,6 +268,11 @@ class TimeSeriesExtractor(Step):
         kwargs.update(date_kwargs)
         where_conds.extend(date_conds)
         if extra_where_conditions:
+            extra_where_conditions = [
+                cond.replace("%", "%%")
+                for cond in extra_where_conditions
+                if "%" in cond
+            ]
             where_conds.extend(extra_where_conditions)
         if where_conds:
             placeholders["where"] = where_conds  # type: ignore